requirements.txt
chatterbox-tts
fastapi
uvicorn[standard]
soundfile
pydantic
